# backend-app/swagger.json
{
    "components": {
...
        "schemas": {
            "Assertion": {
                "properties": {
                    "to": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "be": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "been": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "is": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "that": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "which": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "and": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "has": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "have": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "with": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "at": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "of": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "same": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "but": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "does": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "above": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "gt": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "greaterThan": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "least": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "gte": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "greaterThanOrEqual": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "below": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "lt": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "lessThan": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "most": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "lte": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "lessThanOrEqual": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "instanceof": {
                        "$ref": "#/components/schemas/InstanceOf"
                    },
                    "instanceOf": {
                        "$ref": "#/components/schemas/InstanceOf"
                    },
                    "not": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "deep": {
                        "$ref": "#/components/schemas/Deep"
                    },
                    "ordered": {
                        "$ref": "#/components/schemas/Ordered"
                    },
                    "nested": {
                        "$ref": "#/components/schemas/Nested"
                    },
                    "own": {
                        "$ref": "#/components/schemas/Own"
                    },
                    "any": {
                        "$ref": "#/components/schemas/KeyFilter"
                    },
                    "all": {
                        "$ref": "#/components/schemas/KeyFilter"
                    },
                    "a": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "an": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "include": {
                        "$ref": "#/components/schemas/Include"
                    },
                    "includes": {
                        "$ref": "#/components/schemas/Include"
                    },
                    "contain": {
                        "$ref": "#/components/schemas/Include"
                    },
                    "contains": {
                        "$ref": "#/components/schemas/Include"
                    },
                    "ok": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "true": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "false": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "null": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "undefined": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "NaN": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "exist": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "empty": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "arguments": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "Arguments": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "finite": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "equal": {
                        "$ref": "#/components/schemas/Equal"
                    },
                    "equals": {
                        "$ref": "#/components/schemas/Equal"
                    },
                    "eq": {
                        "$ref": "#/components/schemas/Equal"
                    },
                    "eql": {
                        "$ref": "#/components/schemas/Equal"
                    },
                    "eqls": {
                        "$ref": "#/components/schemas/Equal"
                    },
                    "property": {
                        "$ref": "#/components/schemas/Property"
                    },
                    "ownProperty": {
                        "$ref": "#/components/schemas/Property"
                    },
                    "haveOwnProperty": {
                        "$ref": "#/components/schemas/Property"
                    },
                    "ownPropertyDescriptor": {
                        "$ref": "#/components/schemas/OwnPropertyDescriptor"
                    },
                    "haveOwnPropertyDescriptor": {
                        "$ref": "#/components/schemas/OwnPropertyDescriptor"
                    },
                    "length": {
                        "$ref": "#/components/schemas/Length"
                    },
                    "lengthOf": {
                        "$ref": "#/components/schemas/Length"
                    },
                    "match": {
                        "$ref": "#/components/schemas/Match"
                    },
                    "matches": {
                        "$ref": "#/components/schemas/Match"
                    },
                    "keys": {
                        "$ref": "#/components/schemas/Keys"
                    },
                    "throw": {
                        "$ref": "#/components/schemas/Throw"
                    },
                    "throws": {
                        "$ref": "#/components/schemas/Throw"
                    },
                    "Throw": {
                        "$ref": "#/components/schemas/Throw"
                    },
                    "respondTo": {
                        "$ref": "#/components/schemas/RespondTo"
                    },
                    "respondsTo": {
                        "$ref": "#/components/schemas/RespondTo"
                    },
                    "itself": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "satisfy": {
                        "$ref": "#/components/schemas/Satisfy"
                    },
                    "satisfies": {
                        "$ref": "#/components/schemas/Satisfy"
                    },
                    "closeTo": {
                        "$ref": "#/components/schemas/CloseTo"
                    },
                    "approximately": {
                        "$ref": "#/components/schemas/CloseTo"
                    },
                    "members": {
                        "$ref": "#/components/schemas/Members"
                    },
                    "increase": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "increases": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "decrease": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "decreases": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "change": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "changes": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "extensible": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "sealed": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "frozen": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "oneOf": {
                        "$ref": "#/components/schemas/OneOf"
                    }
                },
                "required": [
                    "to",
                    "be",
                    "been",
                    "is",
                    "that",
                    "which",
                    "and",
                    "has",
                    "have",
                    "with",
                    "at",
                    "of",
                    "same",
                    "but",
                    "does",
                    "above",
                    "gt",
                    "greaterThan",
                    "least",
                    "gte",
                    "greaterThanOrEqual",
                    "below",
                    "lt",
                    "lessThan",
                    "most",
                    "lte",
                    "lessThanOrEqual",
                    "instanceof",
                    "instanceOf",
                    "not",
                    "deep",
                    "ordered",
                    "nested",
                    "own",
                    "any",
                    "all",
                    "a",
                    "an",
                    "include",
                    "includes",
                    "contain",
                    "contains",
                    "ok",
                    "true",
                    "false",
                    "null",
                    "undefined",
                    "NaN",
                    "exist",
                    "empty",
                    "arguments",
                    "Arguments",
                    "finite",
                    "equal",
                    "equals",
                    "eq",
                    "eql",
                    "eqls",
                    "property",
                    "ownProperty",
                    "haveOwnProperty",
                    "ownPropertyDescriptor",
                    "haveOwnPropertyDescriptor",
                    "length",
                    "lengthOf",
                    "match",
                    "matches",
                    "keys",
                    "throw",
                    "throws",
                    "Throw",
                    "respondTo",
                    "respondsTo",
                    "itself",
                    "satisfy",
                    "satisfies",
                    "closeTo",
                    "approximately",
                    "members",
                    "increase",
                    "increases",
                    "decrease",
                    "decreases",
                    "change",
                    "changes",
                    "extensible",
                    "sealed",
                    "frozen",
                    "oneOf"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "NumberComparer": {
                "properties": {},
                "type": "object",
                "additionalProperties": false
            },
...
            "InstanceOf": {
                "properties": {},
                "type": "object",
                "additionalProperties": false
            },
            "Chai.Assertion": {
                "properties": {
                    "to": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "be": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "been": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "is": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "that": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "which": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "and": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "has": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "have": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "with": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "at": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "of": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "same": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "but": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "does": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "above": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "gt": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "greaterThan": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "least": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "gte": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "greaterThanOrEqual": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "below": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "lt": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "lessThan": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "most": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "lte": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "lessThanOrEqual": {
                        "$ref": "#/components/schemas/NumberComparer"
                    },
                    "instanceof": {
                        "$ref": "#/components/schemas/InstanceOf"
                    },
                    "instanceOf": {
                        "$ref": "#/components/schemas/InstanceOf"
                    },
                    "not": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "deep": {
                        "$ref": "#/components/schemas/Deep"
                    },
                    "ordered": {
                        "$ref": "#/components/schemas/Ordered"
                    },
                    "nested": {
                        "$ref": "#/components/schemas/Nested"
                    },
                    "own": {
                        "$ref": "#/components/schemas/Own"
                    },
                    "any": {
                        "$ref": "#/components/schemas/KeyFilter"
                    },
                    "all": {
                        "$ref": "#/components/schemas/KeyFilter"
                    },
                    "a": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "an": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "include": {
                        "$ref": "#/components/schemas/Include"
                    },
                    "includes": {
                        "$ref": "#/components/schemas/Include"
                    },
                    "contain": {
                        "$ref": "#/components/schemas/Include"
                    },
                    "contains": {
                        "$ref": "#/components/schemas/Include"
                    },
                    "ok": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "true": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "false": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "null": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "undefined": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "NaN": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "exist": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "empty": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "arguments": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "Arguments": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "finite": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "equal": {
                        "$ref": "#/components/schemas/Equal"
                    },
                    "equals": {
                        "$ref": "#/components/schemas/Equal"
                    },
                    "eq": {
                        "$ref": "#/components/schemas/Equal"
                    },
                    "eql": {
                        "$ref": "#/components/schemas/Equal"
                    },
                    "eqls": {
                        "$ref": "#/components/schemas/Equal"
                    },
                    "property": {
                        "$ref": "#/components/schemas/Property"
                    },
                    "ownProperty": {
                        "$ref": "#/components/schemas/Property"
                    },
                    "haveOwnProperty": {
                        "$ref": "#/components/schemas/Property"
                    },
                    "ownPropertyDescriptor": {
                        "$ref": "#/components/schemas/OwnPropertyDescriptor"
                    },
                    "haveOwnPropertyDescriptor": {
                        "$ref": "#/components/schemas/OwnPropertyDescriptor"
                    },
                    "length": {
                        "$ref": "#/components/schemas/Length"
                    },
                    "lengthOf": {
                        "$ref": "#/components/schemas/Length"
                    },
                    "match": {
                        "$ref": "#/components/schemas/Match"
                    },
                    "matches": {
                        "$ref": "#/components/schemas/Match"
                    },
                    "keys": {
                        "$ref": "#/components/schemas/Keys"
                    },
                    "throw": {
                        "$ref": "#/components/schemas/Throw"
                    },
                    "throws": {
                        "$ref": "#/components/schemas/Throw"
                    },
                    "Throw": {
                        "$ref": "#/components/schemas/Throw"
                    },
                    "respondTo": {
                        "$ref": "#/components/schemas/RespondTo"
                    },
                    "respondsTo": {
                        "$ref": "#/components/schemas/RespondTo"
                    },
                    "itself": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "satisfy": {
                        "$ref": "#/components/schemas/Satisfy"
                    },
                    "satisfies": {
                        "$ref": "#/components/schemas/Satisfy"
                    },
                    "closeTo": {
                        "$ref": "#/components/schemas/CloseTo"
                    },
                    "approximately": {
                        "$ref": "#/components/schemas/CloseTo"
                    },
                    "members": {
                        "$ref": "#/components/schemas/Members"
                    },
                    "increase": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "increases": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "decrease": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "decreases": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "change": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "changes": {
                        "$ref": "#/components/schemas/PropertyChange"
                    },
                    "extensible": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "sealed": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "frozen": {
                        "$ref": "#/components/schemas/Assertion"
                    },
                    "oneOf": {
                        "$ref": "#/components/schemas/OneOf"
                    }
                },
                "required": [
                    "to",
                    "be",
                    "been",
                    "is",
                    "that",
                    "which",
                    "and",
                    "has",
                    "have",
                    "with",
                    "at",
                    "of",
                    "same",
                    "but",
                    "does",
                    "above",
                    "gt",
                    "greaterThan",
                    "least",
                    "gte",
                    "greaterThanOrEqual",
                    "below",
                    "lt",
                    "lessThan",
                    "most",
                    "lte",
                    "lessThanOrEqual",
                    "instanceof",
                    "instanceOf",
                    "not",
                    "deep",
                    "ordered",
                    "nested",
                    "own",
                    "any",
                    "all",
                    "a",
                    "an",
                    "include",
                    "includes",
                    "contain",
                    "contains",
                    "ok",
                    "true",
                    "false",
                    "null",
                    "undefined",
                    "NaN",
                    "exist",
                    "empty",
                    "arguments",
                    "Arguments",
                    "finite",
                    "equal",
                    "equals",
                    "eq",
                    "eql",
                    "eqls",
                    "property",
                    "ownProperty",
                    "haveOwnProperty",
                    "ownPropertyDescriptor",
                    "haveOwnPropertyDescriptor",
                    "length",
                    "lengthOf",
                    "match",
                    "matches",
                    "keys",
                    "throw",
                    "throws",
                    "Throw",
                    "respondTo",
                    "respondsTo",
                    "itself",
                    "satisfy",
                    "satisfies",
                    "closeTo",
                    "approximately",
                    "members",
                    "increase",
                    "increases",
                    "decrease",
                    "decreases",
                    "change",
                    "changes",
                    "extensible",
                    "sealed",
                    "frozen",
                    "oneOf"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "Object": {
                "description": "Provides functionality common to all JavaScript objects.",
                "properties": {
                    "should": {
                        "$ref": "#/components/schemas/Chai.Assertion"
                    }
                },
                "required": ["should"],
                "type": "object",
                "additionalProperties": false
            },
            "ValidateErrorJSON": {
                "properties": {
                    "message": {
                        "type": "string",
                        "enum": ["Validation failed"],
                        "nullable": false
                    },
                    "details": {
                        "properties": {},
                        "additionalProperties": {},
                        "type": "object"
                    }
                },
                "required": ["message", "details"],
                "type": "object",
                "additionalProperties": false
            },
            "IUser": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "authorities": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "restrictions": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "roles": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "activationKey": {
                        "type": "string"
                    },
                    "accessRestricted": {
                        "type": "boolean"
                    },
                    "githubOauthAccessToken": {
                        "type": "string"
                    },
                    "resetKey": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted": {
                        "type": "boolean"
                    },
                    "deletedBy": {
                        "type": "string"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "updatedBy": {
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "email",
                    "authorities",
                    "restrictions",
                    "roles",
                    "active",
                    "accessRestricted",
                    "createdAt",
                    "updatedAt",
                    "deleted"
                ],
                "type": "object",
                "additionalProperties": false
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "student-workflow-api",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "NFS-Team"
        }
    },
    "openapi": "3.0.0",
    "paths": {
        "/users/{userId}": {
            "post": {
                "operationId": "GetUser",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidateErrorJSON"
                                }
                            }
                        }
                    }
                },
                "description": "Retrieves the details of an existing user.\r\nSupply the unique user ID from either and receive corresponding user details.",
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IUser"
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}